Rename features.h because many linux systems have an include file called features.h which can complicate makefiles that use a search path.
