+++ lib/parser/basic-parsers.h
@@ -33,7 +33,7 @@
 #include "parse-state.h"
 #include "provenance.h"
 #include "user-state.h"
-#include "../common/features.h"
+#include "../common/flang-features.h"
 #include "../common/idioms.h"
 #include "../common/indirection.h"
 #include <cstring>
