+++ lib/semantics/semantics.h
@@ -16,7 +16,7 @@
 #define FORTRAN_SEMANTICS_SEMANTICS_H_
 
 #include "scope.h"
-#include "../common/features.h"
+#include "../common/flang-features.h"
 #include "../evaluate/common.h"
 #include "../evaluate/intrinsics.h"
 #include "../parser/message.h"
