+++ tools/f18/f18.cc
@@ -15,7 +15,7 @@
 // Temporary Fortran front end driver main program for development scaffolding.
 
 #include "../../lib/common/default-kinds.h"
-#include "../../lib/common/features.h"
+#include "../../lib/common/flang-features.h"
 #include "../../lib/evaluate/expression.h"
 #include "../../lib/parser/characters.h"
 #include "../../lib/parser/dump-parse-tree.h"
